requirements.txt
flask==2.2.5
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2
numpy==1.25.2
mysql-connector-python==8.3.0
ace-tools
gunicorn
python-dotenv==1.1.1
flask-cors==6.0.1